update README
 This is a new model and it does not have mature technology
therefore the user experience is not really excellent. -> this is a new model without mature technologies. The user experience is not really excellent.
The cohesion of this sentence is strong so "therefore" is not needed.

Currently, HTML5 and Cordova play an important role in mobile application development because they make anularJS
and ionic technology available on mobile devices. 
-> HTML5 and Cordova are important in mobile application development because they support anularJS and ionic technology on mobile devices.
(HTML5 and Cordova cannot "play". Avoid using void verb such as "make")